Actions refactoring
