+++ src/main/resources/META-INF/plugin.xml
@@ -88,22 +88,22 @@
     -->
 
     <actions>
-        <group id="ProjectViewActionGroup" class="org.vorpal.research.kex.plugin.actions.ProjectViewActionGroup"
+        <group id="ProjectViewActionGroup" class="org.vorpal.research.kex.plugin.actions.projectView.ProjectViewActionGroup"
                popup="true"
                icon="/icons/kex16.svg" text="Kex">
             <add-to-group group-id="ProjectViewCompileGroup"/>
-            <action id="Kex.Actions.ProjectViewAction" class="org.vorpal.research.kex.plugin.actions.ProjectViewAction"
+            <action id="Kex.Actions.ProjectViewAction" class="org.vorpal.research.kex.plugin.actions.projectView.ProjectViewAction"
                     text="Generate Tests"/>
         </group>
 
-        <group id="EditorActionGroup" class="org.vorpal.research.kex.plugin.actions.EditorActionGroup"
+        <group id="EditorActionGroup" class="org.vorpal.research.kex.plugin.actions.editor.EditorActionGroup"
                popup="true"
                icon="/icons/kex16.svg" text="Kex">
             <add-to-group group-id="EditorLangPopupMenu" anchor="after" relative-to-action="Generate"/>
-            <action id="Kex.Actions.EditorAction" class="org.vorpal.research.kex.plugin.actions.EditorAction"
+            <action id="Kex.Actions.EditorAction" class="org.vorpal.research.kex.plugin.actions.editor.EditorAction"
                     text="Generate Tests"/>
             <action id="Kex.Actions.EditorMethodGraphAction"
-                    class="org.vorpal.research.kex.plugin.actions.EditorMethodGraphAction"
+                    class="org.vorpal.research.kex.plugin.actions.editor.EditorMethodGraphAction"
                     text="Graph"/>
         </group>
     </actions>
